% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqltargets-option.R
\name{sqltargets_option_get}
\alias{sqltargets_option_get}
\alias{sqltargets_option_set}
\title{Get or Set sqltargets Options}
\usage{
sqltargets_option_get(option_name)

sqltargets_option_set(option_name, option_value)
}
\arguments{
\item{option_name}{Character. Option name. See Details.}

\item{option_value}{Value to assign to option `x`.}
}
\value{
No return value, called for side effects
}
\description{
Get or Set sqltargets Options
}
\details{
## Available Options

 - `"sqltargets.jinja_block_open"` - character. Length 1. The opening delimiter passed to `jinjar::jinjar_config()`.
 - `"sqltargets.jinja_block_close"` - character. Length 1. The closing delimiter passed to `jinjar::jinjar_config()`.
 - `"sqltargets.jinja_variable_open"` - character. Length 1. The closing delimiter passed to `jinjar::jinjar_config()`.
 - `"sqltargets.jinja_variable_close"` - character. Length 1. The closing delimiter passed to `jinjar::jinjar_config()`.
 - `"sqltargets.jinja_comment_open"` - character. Length 1. The closing delimiter passed to `jinjar::jinjar_config()`.
 - `"sqltargets.jinja_comment_close"` - character. Length 1. The closing delimiter passed to `jinjar::jinjar_config()`.
}
