% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_sql_raw.R
\name{tar_sql_exec}
\alias{tar_sql_exec}
\title{Execute a SQL query.}
\usage{
tar_sql_exec(args, deps)
}
\arguments{
\item{args}{A named list of arguments to `glue::glue_sql()`.}

\item{deps}{An unnamed list of target dependencies of the R Markdown
report, automatically created by `tar_sql_deps()`.}
}
\value{
a data frame.
}
\description{
Internal function needed for `tar_sql()`.
  Users should not invoke it directly.
}
\keyword{internal}
