% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_sql_raw.R
\name{tar_sql_file}
\alias{tar_sql_file}
\title{Define upstream SQL file dep.}
\usage{
tar_sql_file(args)
}
\arguments{
\item{args}{Passed to `fs::path_rel()`.}
}
\value{
A relative file path.
}
\description{
Internal function needed for `tar_sql()`.
  Users should not invoke it directly.
}
\keyword{internal}
