% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqltargets-option.R
\name{sqltargets_option_get}
\alias{sqltargets_option_get}
\alias{sqltargets_option_set}
\title{Get or Set sqltargets Options}
\usage{
sqltargets_option_get(option_name)

sqltargets_option_set(option_name, option_value)
}
\arguments{
\item{option_name}{Character. Option name. See Details.}

\item{option_value}{Value to assign to option `x`.}
}
\value{
No return value, called for side effects
}
\description{
Get or Set sqltargets Options
}
\details{
## Available Options

 - `"sqltargets.target_file_suffix"` - character. Length 1. Suffix appended to target name for SQL file dependency
}
