% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqltargets.R
\docType{package}
\name{sqltargets}
\alias{sqltargets}
\title{\code{sqltargets} package}
\description{
targets extension for SQL files
}
\details{
See the README on
\href{https://github.com/daranzolin/sqltargets#readme}{GitHub}
}
